test-automation/requirements.txt
azure-functions==1.21.3
azure-functions-durable==1.2.9
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.1
openai==1.58.1
pyodbc==5.2.0
requests==2.32.3
openpyxl==3.1.5
